+++ packages/expo-dev-launcher/bundle/screens/HomeScreen.tsx
@@ -1,4 +1,3 @@
-import { useFocusEffect } from '@react-navigation/core';
 import {
   Heading,
   Text,
@@ -38,12 +37,18 @@ import { useUpdatesConfig } from '../providers/UpdatesConfigProvider';
 import { DevSession } from '../types';
 
 export type HomeScreenProps = {
+  fetchOnMount?: boolean;
   pollInterval?: number;
   pollAmount?: number;
   navigation?: any;
 };
 
-export function HomeScreen({ pollInterval = 1000, pollAmount = 5, navigation }: HomeScreenProps) {
+export function HomeScreen({
+  fetchOnMount = false,
+  pollInterval = 1000,
+  pollAmount = 5,
+  navigation,
+}: HomeScreenProps) {
   const modalStack = useModalStack();
   const [inputValue, setInputValue] = React.useState('');
   const [loadingUrl, setLoadingUrl] = React.useState('');
@@ -54,14 +59,13 @@ export function HomeScreen({ pollInterval = 1000, pollAmount = 5, navigation }: 
 
   const crashReport = useCrashReport();
 
-  const hasDevSessions = devSessions?.length > 0;
-  useFocusEffect(
-    React.useCallback(() => {
-      if (!hasDevSessions) {
-        pollAsync({ pollAmount, pollInterval });
-      }
-    }, [hasDevSessions, pollAmount, pollInterval])
-  );
+  const initialDevSessionData = React.useRef(devSessions);
+
+  React.useEffect(() => {
+    if (initialDevSessionData.current.length === 0 && fetchOnMount) {
+      pollAsync({ pollAmount, pollInterval });
+    }
+  }, [fetchOnMount, pollInterval, pollAmount, pollAsync]);
 
   const onLoadUrl = async (url: string) => {
     setLoadingUrl(url);
@@ -137,7 +141,7 @@ export function HomeScreen({ pollInterval = 1000, pollAmount = 5, navigation }: 
 
                 <Spacer.Horizontal />
 
-                {devSessions?.length > 0 && (
+                {devSessions.length > 0 && (
                   <Button.FadeOnPressContainer
                     bg="ghost"
                     rounded="full"
@@ -153,7 +157,7 @@ export function HomeScreen({ pollInterval = 1000, pollAmount = 5, navigation }: 
 
               <View px="medium">
                 <View>
-                  {devSessions?.length === 0 && (
+                  {devSessions.length === 0 && (
                     <>
                       <View padding="medium" bg="default" roundedTop="large">
                         <Text>Start a local development server with:</Text>